+++ src/Public/Pages.ps1
@@ -174,6 +174,7 @@ function Set-PodeWebLoginPage
             Page = @{
                 Name = 'Home'
                 Title = 'Home'
+                DisplayName = 'Home'
                 Path = '/'
                 IsSystem = $true
             }
@@ -205,6 +206,10 @@ function Set-PodeWebHomePage
 
         [Parameter()]
         [string]
+        $DisplayName,
+
+        [Parameter()]
+        [string]
         $Title,
 
         [Parameter()]
@@ -229,8 +234,12 @@ function Set-PodeWebHomePage
     }
 
     # set page title
+    if ([string]::IsNullOrWhiteSpace($DisplayName)) {
+        $DisplayName = 'Home'
+    }
+
     if ([string]::IsNullOrWhiteSpace($Title)) {
-        $Title = 'Home'
+        $Title = $DisplayName
     }
 
     # route path
@@ -242,7 +251,8 @@ function Set-PodeWebHomePage
         ObjectType = 'Page'
         Path = $routePath
         Name = 'Home'
-        Title = $Title
+        Title = [System.Net.WebUtility]::HtmlEncode($Title)
+        DisplayName = [System.Net.WebUtility]::HtmlEncode($DisplayName)
         NoTitle = $NoTitle.IsPresent
         Navigation = $Navigation
         Layouts = $Layouts
@@ -316,6 +326,10 @@ function Add-PodeWebPage
         [Parameter(Mandatory=$true)]
         [string]
         $Name,
+
+        [Parameter()]
+        [string]
+        $DisplayName,
 
         [Parameter()]
         [string]
@@ -400,8 +414,12 @@ function Add-PodeWebPage
     }
 
     # set page title
+    if ([string]::IsNullOrWhiteSpace($DisplayName)) {
+        $DisplayName = $Name
+    }
+
     if ([string]::IsNullOrWhiteSpace($Title)) {
-        $Title = $Name
+        $Title = $DisplayName
     }
 
     # build the route path
@@ -414,6 +432,7 @@ function Add-PodeWebPage
         Path = $routePath
         Name = $Name
         Title = [System.Net.WebUtility]::HtmlEncode($Title)
+        DisplayName = [System.Net.WebUtility]::HtmlEncode($DisplayName)
         NoTitle = $NoTitle.IsPresent
         NoBackArrow = $NoBackArrow.IsPresent
         NoBreadcrumb = $NoBreadcrumb.IsPresent
@@ -515,6 +534,7 @@ function Add-PodeWebPage
             Write-PodeWebViewResponse -Path 'index' -Data @{
                 Page = $global:PageData
                 Title = $global:PageData.Title
+                DisplayName = $global:PageData.DisplayName
                 Theme = $theme
                 Navigation = $navigation
                 Breadcrumb = $breadcrumb
@@ -579,7 +599,7 @@ function Add-PodeWebPageLink
 
         [Parameter()]
         [string]
-        $Title,
+        $DisplayName,
 
         [Parameter()]
         [string]
@@ -633,8 +653,8 @@ function Add-PodeWebPageLink
     }
 
     # set page title
-    if ([string]::IsNullOrWhiteSpace($Title)) {
-        $Title = $Name
+    if ([string]::IsNullOrWhiteSpace($DisplayName)) {
+        $DisplayName = $Name
     }
 
     # setup page meta
@@ -642,7 +662,7 @@ function Add-PodeWebPageLink
         ComponentType = 'Page'
         ObjectType = 'Link'
         Name = $Name
-        Title = [System.Net.WebUtility]::HtmlEncode($Title)
+        DisplayName = [System.Net.WebUtility]::HtmlEncode($DisplayName)
         NewTab = $NewTab.IsPresent
         Icon = $Icon
         Group = $Group